user actions component fixes
