--- src/app/components/user-actions/user-actions.component.html
@@ -1,16 +1,0 @@
-<table class="table table-borderless table-striped">
-    <thead>
-        <tr>
-            <th>Action</th>
-            <th>Token Change</th>
-        </tr>
-    </thead>
-    <tbody>
-        <ng-container *ngFor="let action of userActions">
-            <tr>
-                <td>{{action.description}}</td>
-                <td>{{action.token_change}}</td>
-            </tr>
-        </ng-container>
-    </tbody>
-</table>
